Reduced duplication between InteractiveNav and Nav.
